# oka_template/ooka_task/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation ",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/SuratMasuk": {
            "get": {
                "tags": [
                    "SuratMasuk"
                ],
                "summary": "Get all Data suratMasuk",
                "description": "",
                "operationId": "index.suratMasuk",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratMasuk/store": {
            "post": {
                "tags": [
                    "SuratMasuk"
                ],
                "summary": "Tambah Data suratMasuk",
                "description": "",
                "operationId": "store.suratMasuk",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "name field : asal_surat, no_surat, tujuan_surat, status('Rahasia', 'Penting', 'Segera', 'Biasa'), file_surat(file)",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "asal_surat": {
                                            "type": "string",
                                            "example": "Lurah "
                                        },
                                        "no_surat": {
                                            "type": "string",
                                            "example": "12/3213/213"
                                        },
                                        "tujuan_surat": {
                                            "type": "string",
                                            "example": "Malino"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "'Rahasia', 'Penting', 'Segera', 'Biasa'"
                                        },
                                        "file_surat": {
                                            "type": "string",
                                            "example": "file.pdf,docx"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratMasuk/": {
            "get": {
                "tags": [
                    "SuratMasuk"
                ],
                "summary": "Edit Data suratMasuk",
                "description": "tambah /id",
                "operationId": "edit.suratMasuk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratMasuk/update": {
            "put": {
                "tags": [
                    "SuratMasuk"
                ],
                "summary": "Update Data suratMasuk",
                "description": "kirim id lewat form",
                "operationId": "update.suratMasuk",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "name field : id, asal_surat, no_surat, tujuan_surat, status('Rahasia', 'Penting', 'Segera', 'Biasa'), file_surat(file)",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "asal_surat": {
                                            "type": "string",
                                            "example": "Lurah "
                                        },
                                        "no_surat": {
                                            "type": "string",
                                            "example": "12/3213/213"
                                        },
                                        "tujuan_surat": {
                                            "type": "string",
                                            "example": "Malino"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "'Rahasia', 'Penting', 'Segera', 'Biasa'"
                                        },
                                        "file_surat": {
                                            "type": "string",
                                            "example": "file.pdf,docx"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratMasuk/destroy": {
            "delete": {
                "tags": [
                    "SuratMasuk"
                ],
                "summary": "Hapus Data suratMasuk",
                "description": "tambah /id",
                "operationId": "destroy.suratMasuk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "name field : id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        }
    }
}